+++ lib/datasource/local/account_service.dart
@@ -1,41 +1,5 @@
-import 'dart:convert';
-
-import 'package:equatable/equatable.dart';
+import 'package:seeds/datasource/local/models/account.dart';
 import 'package:seeds/datasource/local/settings_storage.dart';
-
-class Account extends Equatable {
-  String address;
-  String name;
-  Account(this.address, this.name);
-
-  Map<String, dynamic> toJson() {
-    return {
-      "address": address,
-      "name": name,
-    };
-  }
-
-  factory Account.fromJson(Map<String, dynamic> json) {
-    return Account(
-      json["address"],
-      json["name"],
-    );
-  }
-
-  static List<Account> listFromJson(String jsonString) {
-    final List items = json.decode(jsonString) as List;
-    return items.map((e) => Account.fromJson(e)).toList();
-  }
-
-  static String jsonFromList(List<Account> accounts) {
-    final res = json.encode(accounts);
-    print("json str: $res");
-    return res;
-  }
-
-  @override
-  List<Object?> get props => [name, address];
-}
 
 class AccountService {
   Future<List<Account>> loadAccounts() async {
@@ -53,7 +17,7 @@ class AccountService {
     if (privateKey.contains(",")) {
       throw ArgumentError("illegal character in private key: ',': $privateKey");
     }
-    Account? result = null;
+    Account? result;
     final public = await publicKeyForPrivateKey(privateKey);
     if (public != null) {
       final account = Account(name, public);
