+++ lib/screens/authentication/import_key/interactor/viewmodels/import_key_bloc.dart
@@ -7,10 +7,8 @@ import 'package:seeds/screens/authentication/import_key/import_key_errors.dart';
 import 'package:seeds/screens/authentication/import_key/interactor/mappers/import_key_state_mapper.dart';
 import 'package:seeds/screens/authentication/import_key/interactor/usecases/check_private_key_use_case.dart';
 import 'package:seeds/screens/authentication/import_key/interactor/usecases/generate_key_from_recovery_words_use_case.dart';
-import 'package:seeds/screens/authentication/import_key/interactor/usecases/generate_key_from_seeds_passport_words_use_case.dart';
 import 'package:seeds/screens/authentication/import_key/interactor/usecases/import_key_use_case.dart';
 import 'package:seeds/utils/helpers.dart';
-import 'package:seeds/utils/result_extension.dart';
 
 part 'import_key_event.dart';
 
@@ -62,15 +60,11 @@ class ImportKeyBloc extends Bloc<ImportKeyEvent, ImportKeyState> {
   Future<void> _findAccountByKey(FindAccountByKey event, Emitter<ImportKeyState> emit) async {
     emit(state.copyWith(pageState: PageState.loading));
     final publicKey = CheckPrivateKeyUseCase().isKeyValid(event.privateKey);
-    final alternatePublicKey =
-        event.alternatePrivateKey != null ? CheckPrivateKeyUseCase().isKeyValid(event.alternatePrivateKey!) : null;
 
     if (publicKey == null || publicKey.isEmpty) {
       emit(state.copyWith(pageState: PageState.failure, error: ImportKeyError.invalidPrivateKey));
     } else {
       final results = await ImportKeyUseCase().run(publicKey);
-      final List<Result> alternateResults =
-          alternatePublicKey != null ? await ImportKeyUseCase().run(alternatePublicKey) : [];
 
       emit(
         ImportKeyStateMapper().mapResultsToState(
