+++ lib/screens/profile_screens/recovery_phrase/interactor/viewmodels/recovery_phrase_bloc.dart
@@ -1,6 +1,5 @@
 import 'package:bloc/bloc.dart';
 import 'package:seeds/datasource/local/models/auth_data_model.dart';
-import 'package:seeds/datasource/local/settings_storage.dart';
 
 part 'recovery_phrase_event.dart';
 part 'recovery_phrase_state.dart';
