+++ test/account_service_test.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Test something', () {});
+}
